Added search endpoint scaffolding
