+++ .babelrc
@@ -1,12 +1,3 @@
 {
-    "presets": [
-        [
-            "env",
-            {
-                "targets": {
-                    "node": true
-                }
-            }
-        ]
-    ]
+    "presets": ["@babel/preset-env"]
 }
