+++ .gitignore
@@ -1,3 +1,4 @@
+.vscode
 # Logs
 logs
 *.log
@@ -102,3 +103,5 @@ dist
 
 # TernJS port file
 .tern-port
+
+yarn.lock
