--- routes/api/users.js
@@ -1,1 +1,0 @@
-console.log('object');
