--- routes/app.js
@@ -1,6 +1,0 @@
-import express from 'express';
-
-const app = express();
-
-
-export default app;
