+++ src/api/routes/index.js
@@ -1,0 +1,9 @@
+import express from 'express';
+
+import searchRouter from '../routes/search.router';
+
+const api = express();
+
+api.use('/search', searchRouter)
+
+export default api;
